Move all the arch plugins to the new plugin structure structure
