+++ libr/arch/meson.build
@@ -6,17 +6,18 @@ r_arch_sources = [
   'arch_session.c',
   'arch_cond.c',
   'arch_value.c',
-  'p/arch_null.c',
-  'p/arch_i4004.c',
-  'p/arch_amd29k.c',
-  'p/arch_jdh8.c',
+  'p/null/plugin.c',
+  'p/any_as/plugin.c',
+  'p/i4004/plugin.c',
+  'p/amd29k/plugin.c',
+  'p/jdh8/plugin.c',
   'p/pickle/plugin.c',
-  'p/arch_rsp.c',
+  'p/rsp/plugin.c',
   'p/rsp/rsp_idec.c',
   'p/riscv/plugin.c',
-  'p/arch_sh.c',
+  'p/sh/plugin.c',
   'p/sh/gnu/sh-dis.c',
-  'p/arch_v810.c',
+  'p/v810/plugin.c',
   'p/v810/v810_disas.c',
 ]
